Create llist_api.py
Started implementation for the llist_api.py. As of right now, it is a copy of the avl api "api.py" Some stuff will be kept the same, but some stuff like the action function, will be changed. Also need to wait on the completion of llist_wrapper.py so that I can use it within this and utils.py