Create tests-temp-josh.py
This test will work if I configure the django urls to have the /temp-josh/spawn_ant/<str::id>/ URL. I have done this and the tests pass. However, I changed the URL configuration back to normal since someone else is working with that file right now. This spawn ant functionality and tests for it will be configured when I talk to David